% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel.R
\name{fuel_context}
\alias{fuel_context}
\title{Show FuelEconomy.gov package context for LLM use}
\usage{
fuel_context()
}
\value{
Invisible string with full context
}
\description{
Show FuelEconomy.gov package context for LLM use
}
