% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel.R
\name{fuel_makes}
\alias{fuel_makes}
\title{Get makes available for a given year}
\usage{
fuel_makes(year)
}
\arguments{
\item{year}{Model year (e.g. 2024)}
}
\value{
tibble: text (make name), value (make name)
}
\description{
Get makes available for a given year
}
