% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel.R
\name{fuel_models}
\alias{fuel_models}
\title{Get models available for a given year and make}
\usage{
fuel_models(year, make)
}
\arguments{
\item{year}{Model year (e.g. 2024)}

\item{make}{Make name (e.g. "Toyota", "Ford")}
}
\value{
tibble: text (model name), value (model id)
}
\description{
Get models available for a given year and make
}
