% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel.R
\name{fuel_vehicle}
\alias{fuel_vehicle}
\title{Get vehicle details by ID}
\usage{
fuel_vehicle(id)
}
\arguments{
\item{id}{Vehicle ID from fuel_models()}
}
\value{
tibble with one row: id, year, make, model, trany, drive, cylinders,
displ, fuel_type, city_mpg, highway_mpg, comb_mpg, co2
}
\description{
Get vehicle details by ID
}
