% Generated by build script
\name{fuel_vehicle}
\alias{fuel_vehicle}
\title{Get vehicle details by ID}
\usage{
fuel_vehicle()
}
\arguments{
\item{id}{Vehicle ID from fuel_models()}
}
\value{tibble with one row: id, year, make, model, trany, drive, cylinders,}
\description{Get vehicle details by ID}
